refactor: unused imports and style
